requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-modeltranslation==0.19.14
idna==3.10
pillow==11.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
